refactor(pushover): move configuration types into pushover package
