+++ config/notifiers_test.go
@@ -193,98 +193,6 @@ custom_fields:
 	}
 }
 
-func TestPushoverUserKeyIsPresent(t *testing.T) {
-	in := `
-user_key: ''
-`
-	var cfg PushoverConfig
-	err := yaml.UnmarshalStrict([]byte(in), &cfg)
-
-	expected := "one of user_key or user_key_file must be configured"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%v", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%v\ngot:\n%v", expected, err.Error())
-	}
-}
-
-func TestPushoverUserKeyOrUserKeyFile(t *testing.T) {
-	in := `
-user_key: 'user key'
-user_key_file: /pushover/user_key
-`
-	var cfg PushoverConfig
-	err := yaml.UnmarshalStrict([]byte(in), &cfg)
-
-	expected := "at most one of user_key & user_key_file must be configured"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%v", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%v\ngot:\n%v", expected, err.Error())
-	}
-}
-
-func TestPushoverTokenIsPresent(t *testing.T) {
-	in := `
-user_key: '<user_key>'
-token: ''
-`
-	var cfg PushoverConfig
-	err := yaml.UnmarshalStrict([]byte(in), &cfg)
-
-	expected := "one of token or token_file must be configured"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%v", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%v\ngot:\n%v", expected, err.Error())
-	}
-}
-
-func TestPushoverTokenOrTokenFile(t *testing.T) {
-	in := `
-token: 'pushover token'
-token_file: /pushover/token
-user_key: 'user key'
-`
-	var cfg PushoverConfig
-	err := yaml.UnmarshalStrict([]byte(in), &cfg)
-
-	expected := "at most one of token & token_file must be configured"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%v", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%v\ngot:\n%v", expected, err.Error())
-	}
-}
-
-func TestPushoverHTMLOrMonospace(t *testing.T) {
-	in := `
-token: 'pushover token'
-user_key: 'user key'
-html: true
-monospace: true
-`
-	var cfg PushoverConfig
-	err := yaml.UnmarshalStrict([]byte(in), &cfg)
-
-	expected := "at most one of monospace & html must be configured"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%v", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%v\ngot:\n%v", expected, err.Error())
-	}
-}
-
 func TestLoadSlackConfiguration(t *testing.T) {
 	tests := []struct {
 		in       string
