+++ notify/pushover/pushover.go
@@ -25,7 +25,6 @@ import (
 
 	commoncfg "github.com/prometheus/common/config"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/template"
 	"github.com/prometheus/alertmanager/types"
@@ -42,7 +41,7 @@ const (
 
 // Notifier implements a Notifier for Pushover notifications.
 type Notifier struct {
-	conf    *config.PushoverConfig
+	conf    *PushoverConfig
 	tmpl    *template.Template
 	logger  *slog.Logger
 	client  *http.Client
@@ -51,7 +50,7 @@ type Notifier struct {
 }
 
 // New returns a new Pushover notifier.
-func New(c *config.PushoverConfig, t *template.Template, l *slog.Logger, httpOpts ...commoncfg.HTTPClientOption) (*Notifier, error) {
+func New(c *PushoverConfig, t *template.Template, l *slog.Logger, httpOpts ...commoncfg.HTTPClientOption) (*Notifier, error) {
 	client, err := notify.NewClientWithTracing(*c.HTTPConfig, "pushover", httpOpts...)
 	if err != nil {
 		return nil, err
